+++ HISTORY.rst
@@ -3,6 +3,8 @@ Changelog
 
 0.4.0 (unreleased)
 ------------------
+- New tokenizer module with ``WordTokenizer`` and ``SentenceTokenizer``. Both textblob and NLTK tokenizer objects and be passed to TextBlob's constructor. Tokens are accessed through the new ``tokens`` property.
+- `Backwards-incompatible`: ``TextBlob.json()`` is now a method, not a property. This allows you to pass arguments (the same that you would pass to ``json.dumps()``).
 - New home for documentation: https://textblob.readthedocs.org/
 - Fix bug with adding blobs to byte strings.